fix: heatmap summary may contain old values
Fixes some issues when reading and writing to the heatmap concurrently.

We combine the subtract and clear steps required to zero out a slice in the
heatmap with a combined function in the histogram crate. This allows us to make
use of a swap + fetch_sub and eliminates a potential case where an increment
could happen between the load + fetch_sub and the eventual store of the two step
subtract + clear.

In the heatmap crate, we create one additional slice per heatmap so that we can
always maintain a cleared histogram in the heatmap to begin writing into. We
also move the bulk of the tick() logic, which is responsible for moving the
pointers forward and clearing the oldest slice, into a critical block. This
should eliminate a lot of potential concurrency bugs.