heatmap: allow recording sample into older slice (#26)
Previously, the heatmap would only record into the "current" slice.
This meant that we could only record past events into the most
recent time slice.

This change adds some additional fields to the heatmap to track the
represented span of time for the overall heatmap and each time slice.
We leave a fast path in-place for when the sample time falls into
the span for the current slice. If it does not, we calculate the
correct offset into the heatmap and record into that older slice.

By making this change, we can correctly record based on event start
time instead of event stop time.