fix: ratelimit bugs
Fixes three open issues:

Users may expect `.wait()` to block on the first call immediately
after creating a `Ratelimiter`. Fixed by setting the first tick
to be in the future so that `.wait()` will block for roughly one
interval. (#33)

Quantum argument did not behave as documented. Fixes the interval
calculation logic so that `quantum` tokens are added at the
configured `rate`. For example, rate 1 and quantum 10 adds 10
tokens every second. Rate 10 and quantum 1 would add 1 token every
100ms. (#35)

Rate limit did not apply after a period of inactivity due to bad
logic in the `tick()` function. Refactored that function to make
the implementation more straightforward and eliminate the bugs.