feat(metriken): rewrite metrics library
A clean reimplementation of the metriken metrics library. That
makes some slightly different design choices. Metrics may now have
associated metadata and an associated formatter. This allows for
adding scoped metrics and to define a particular formatter for each
metric.

Overall this adds a lot of flexibility through a richer metrics
representation.