metriken: replace heatmap with histograms (#77)
The heatmap implementation requires maintaining a buffer of delta'd
histograms. This results in cost on the write path to maintain the
ring buffer and memory costs to maintain an appropriately deep
buffer to cover the desired time span with fine enough granularity
to reduce reporting errors.

The latest histogram crate provides both plain and atomic histogram
implementations that produce snapshots. We can migrate from using
heatmaps to using histograms with periodic or on-demand snapshots
in the consumer. This greatly reduces the cost on the write path
in exchange for higher cost on the reporting path. For typical
metrics cases, this tradeoff should be acceptable.