histogram: add sparse representation
Add a sparse representation of the histogram which only stores non-zero
buckets. Optionally add serde bindings to the SparseHistogram so that
it can be serialized as a storage format. The histogram store's bucket
data in a columnar representation, with the indices and the values stored
in individual arrays, as this results in smaller footprint when serialized.