Backport metriken v0.3 to be a wrapper around metriken-core
This changes metriken to remain (almost) entirely API-compatible while
still being based on the common metriken-core crate.

The sum total of actual breaking changes that are required for this are:
- MetricEntry::metric() now returns a new CoreMetric type instead of
  dyn Metric. This has the exact same API but the type is different.

I would fix this if I could but it is not possible to cast between
traits in rust. As is, I think this breakage is acceptable.