fix: percentile now properly finds the min bucket
Previously, the percentiles functions would report the first
bucket as the 0.0th percentile, even if it was empty. We would
expect instead that it returns the first bucket with a non-zero
count.

This fix wraps the calculation so that low percentiles will always
need to find a bucket with at least a count of one.

As a result, percentiles() can now be used to find the proper min,
max, and any other percentiles one may need to report out.